requirements_airflow.txt
lxml
python-decouple
apache-airflow
apache-airflow-providers-apache-spark
requests
pandas
python-dotenv